Add support for turning off repairReplication as part of the tablet health check if other reparenting events are going on.
There are three kinds of reparenting events.

(1) and (2) PlannedReparenting and EmergencyReparenting

Adds a timestamp of the reparenting event as part of the shard record to check for before
repairReplication actually occurs in both EmergencyReparenting and
PlannedReparenting. Removes the new timestamp in the shard record after
reparenting either finishes successfully or unsuccessfully.

(3) External Reparenting

Adds a new vtctl command called 'PreExternallyReparentShard' to call to
set the additional element on the shard record before beginning external
reparenting.

Resets the new timestamp in the shard record back to 0 as part of
calling TabletExternallyReparent, which is _after_ a reparenting event
has occurred.

Signed-off-by: Maggie Zhou <mzhou@slack-corp.com>